+++ DataAccess.Model/SharedModels/GameContext.cs
@@ -9,7 +9,7 @@ namespace DataAccess.Model.SharedModels
         public Guid GameId { get; set; }
         public Guid RoundId { get; set; }
         public PhaseType CurrentPhase { get; set; }
-        public int RoundNumber { get; set; }
+        public long RoundNumber { get; set; }
         public List<Participant> Participants { get; set; }
     }
 }
